fix/fixvalidators | refactoring all validator classes and fix the classes Services of these validators
