+++ src/main/java/com/challenge/hotel_california/validatorRefactor/IValidatorBookingsDeleteAndCheckout.java
@@ -1,0 +1,7 @@
+package com.challenge.hotel_california.validatorRefactor;
+
+import com.challenge.hotel_california.model.Booking;
+
+public interface IValidatorBookingsDeleteAndCheckout {
+    void verifyBookingsValidators(Long id, Booking bookingFound);
+}
