+++ src/main/java/com/challenge/hotel_california/validatorRefactor/IValidatorCustomers.java
@@ -1,0 +1,8 @@
+package com.challenge.hotel_california.validatorRefactor;
+
+import com.challenge.hotel_california.DTOs.CustomerEntryDTO;
+
+public interface IValidatorCustomers {
+    void verifyCustomersValidators(CustomerEntryDTO customerEntryDTO);
+
+}
